REFACTOR: Modified all IDs to be of type long instead of int

+++ src/main/java/edu/sandiego/comp305/sp24/schoolSim/model/Alumni.java
@@ -9,7 +9,7 @@ public class Alumni extends Person {
     private Date graduationDate;
     private DegreeType degreeType;
 
-    public Alumni(int id) {
+    public Alumni(long id) {
         super(id);
 
         try {
