+++ src/test/java/edu/sandiego/comp305/sp24/schoolSim/model/FacultyTest.java
@@ -82,7 +82,7 @@ public class FacultyTest {
         deleteWithId(faculty.getId(), "Person");
     }
 
-    void deleteWithId(int id, String table) {
+    void deleteWithId(long id, String table) {
         try {
             String sql = "DELETE FROM " + table + " WHERE `id` = ?";
