+++ src/test/java/edu/sandiego/comp305/sp24/schoolSim/model/RoomTest.java
@@ -54,7 +54,7 @@ class RoomTest {
         deleteWithId(room.getId(), "Room");
     }
 
-    void deleteWithId(int id, String table) {
+    void deleteWithId(long id, String table) {
         try {
             String sql = "DELETE FROM " + table + " WHERE `id` = ?";
